[8098] Support uint32 spell ids in code.
* Propertly work with uint32 spell ids in player action bar
* Fix in same time bug with not save equipment set button with id==0
* Merge misc field in character_action and playercreateinfo_action to action field as 3 byte
* Propertly load uint32 spell ids from character_spell
* Fixed types for some pet/creature related structure for spell id storing.

+++ src/game/Pet.h
@@ -69,7 +69,7 @@ enum PetSpellType
 
 struct PetSpell
 {
-    uint16 active;                                          // use instead enum (not good use *uint16* limited enum in case when value in enum not possitive in *int16*)
+    uint8 active;                                           // use instead enum (not good use *uint8* limited enum in case when value in enum not possitive in *int8*)
 
     PetSpellState state : 8;
     PetSpellType type   : 8;
